requirements.txt
# Core dependencies for MASH IoT Device
# Sensor and hardware control
adafruit-circuitpython-scd4x>=1.2.0
RPi.GPIO>=0.7.1

# MQTT communication
paho-mqtt>=1.6.1

# Display and UI
Pillow>=9.0.0
rich>=13.0.0

# HTTP requests for API calls
requests>=2.28.0
urllib3>=1.26.0

# Data handling
numpy>=1.21.0
pandas>=1.4.0

# Configuration and logging
pyyaml>=6.0
python-dotenv>=0.19.0

# Testing
pytest>=7.0.0
pytest-mock>=3.7.0

# Development tools
black>=22.0.0
pylint>=2.13.0
flake8>=4.0.0

# System monitoring
psutil>=5.9.0

# Web server for local API
flask>=2.0.0
flask-restful>=0.3.9
flask-cors>=3.0.10

# Serial communication with Arduino
pyserial>=3.5

# mDNS for local discovery
zeroconf>=0.38.0

# Built-in modules (commented out as reminders)
# sqlite3 (built-in)
# uuid (built-in)
# json (built-in)
# datetime (built-in)
# threading (built-in)
# logging (built-in)
